Move all files to p2p
