+++ p2p/index.js
@@ -11,7 +11,7 @@ const v8 = require('v8');
 const _ = require('lodash');
 const ainUtil = require('@ainblockchain/ain-util');
 const logger = require('../logger')('P2P_SERVER');
-const P2pClient = require('./client');
+const P2pClient = require('../p2p/client');
 const Consensus = require('../consensus');
 const { Block } = require('../blockchain/block');
 const Transaction = require('../tx-pool/transaction');
